requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-ai-generative==1.0.0b8
azure-ai-ml==1.16.1
azure-ai-resources==1.0.0b8
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==22.0.0
azure-storage-blob==12.20.0
azure-storage-file-datalake==12.15.0
azure-storage-file-share==12.16.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
decorator==4.4.2
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
docopt==0.6.2
email_validator==2.2.0
entrypoints==0.4
fastapi==0.111.1
fastapi-cli==0.0.4
fonttools==4.53.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.5.1
importlib-metadata==7.0.0
isodate==0.6.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.6
langchain-core==0.2.11
langchain-openai==0.1.14
langchain-text-splitters==0.2.2
langsmith==0.1.83
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
mdurl==0.1.2
mlflow-skinny==2.13.0
moviepy==1.0.3
msal==1.28.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
num2words==0.5.13
numpy==1.26.4
oauthlib==3.2.2
openai==1.35.10
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
opencv-python==4.10.0.84
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.6
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
plotly==5.22.0
portalocker==2.8.2
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydash==8.0.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
starlette==0.37.2
strictyaml==1.7.3
tenacity==8.3.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.0
